+++ cloudfile/backend/include/cloudfile/api/search_controller.h
@@ -13,7 +13,7 @@ class SearchController : public drogon::HttpController<SearchController> {
 public:
     METHOD_LIST_BEGIN
     ADD_METHOD_TO(SearchController::search, "/api/search", drogon::Get,
-                  "cloudfile::api::SessionAuthFilter");
+                  "cloudfile::api::AuthFilter");
     METHOD_LIST_END
 
     void search(const drogon::HttpRequestPtr& req,